+++ graphix/parameter.py
@@ -334,6 +334,26 @@ else:
 T = TypeVar("T")
 
 
+def check_expression_or_complex(value: object) -> ExpressionOrComplex:
+    """Check that the given object is of type ExpressionOrComplex and return it."""
+    if isinstance(value, Expression):
+        return value
+    if isinstance(value, SupportsComplex):
+        return complex(value)
+    msg = f"ExpressionOrComplex expected, but {type(value)} found."
+    raise TypeError(msg)
+
+
+def check_expression_or_float(value: object) -> ExpressionOrFloat:
+    """Check that the given object is of type ExpressionOrFloat and return it."""
+    if isinstance(value, Expression):
+        return value
+    if isinstance(value, SupportsFloat):
+        return float(value)
+    msg = f"ExpressionOrFloat expected, but {type(value)} found."
+    raise TypeError(msg)
+
+
 @overload
 def subs(value: ExpressionOrFloat, variable: Parameter, substitute: ExpressionOrSupportsFloat) -> ExpressionOrFloat: ...