+++ graphix/sim/density_matrix.py
@@ -6,36 +6,40 @@ Simulate MBQC with density matrix representation.
 from __future__ import annotations
 
 import copy
-import sys
+import dataclasses
+import math
 from collections.abc import Collection, Iterable
-from typing import TYPE_CHECKING, SupportsComplex, SupportsFloat
+from dataclasses import dataclass
+from typing import TYPE_CHECKING, SupportsComplex
 
 import numpy as np
+from typing_extensions import override
 
 from graphix import linalg_validations as lv
-from graphix import parameter, states
+from graphix import parameter
 from graphix.channels import KrausChannel
-from graphix.parameter import Expression, ExpressionOrSupportsComplex
-from graphix.sim.base_backend import Backend, State
+from graphix.parameter import Expression, ExpressionOrFloat, ExpressionOrSupportsComplex
+from graphix.sim.base_backend import DenseState, DenseStateBackend, Matrix, kron, matmul, outer, tensordot, vdot
 from graphix.sim.statevec import CNOT_TENSOR, CZ_TENSOR, SWAP_TENSOR, Statevec
-from graphix.states import BasicStates
+from graphix.states import BasicStates, State
 
 if TYPE_CHECKING:
-    from collections.abc import Mapping
-
-    import numpy.typing as npt
+    from collections.abc import Mapping, Sequence
 
     from graphix.parameter import ExpressionOrSupportsFloat, Parameter
+    from graphix.sim.data import Data
 
 
-class DensityMatrix(State):
+class DensityMatrix(DenseState):
     """DensityMatrix object."""
+
+    rho: Matrix
 
     def __init__(
         self,
         data: Data = BasicStates.PLUS,
         nqubit: int | None = None,
-    ):
+    ) -> None:
         """Initialize density matrix objects.
 
         The behaviour builds on the one of *graphix.statevec.Statevec*.
@@ -61,40 +65,42 @@ class DensityMatrix(State):
         if nqubit is not None and nqubit < 0:
             raise ValueError("nqubit must be a non-negative integer.")
 
-        def check_size_consistency(mat) -> None:
+        def check_size_consistency(mat: Matrix) -> None:
             if nqubit is not None and mat.shape != (2**nqubit, 2**nqubit):
                 raise ValueError(
                     f"Inconsistent parameters between nqubit = {nqubit} and the shape of the provided density matrix = {mat.shape}."
                 )
 
         if isinstance(data, DensityMatrix):
-            check_size_consistency(data)
+            check_size_consistency(data.rho)
             # safe: https://numpy.org/doc/stable/reference/generated/numpy.ndarray.copy.html
             self.rho = data.rho.copy()
             return
         if isinstance(data, Iterable):
             input_list = list(data)
-            if len(input_list) != 0:
-                # needed since Object is iterable but not subscribable!
-                try:
-                    if isinstance(input_list[0], Iterable) and isinstance(
-                        input_list[0][0], (Expression, SupportsComplex, SupportsFloat)
-                    ):
-                        self.rho = np.array(input_list)
-                        if not lv.is_qubitop(self.rho):
-                            raise ValueError("Cannot interpret the provided density matrix as a qubit operator.")
-                        check_size_consistency(self.rho)
-                        if self.rho.dtype != "O":
-                            if not lv.is_unit_trace(self.rho):
-                                raise ValueError("Density matrix must have unit trace.")
-                            if not lv.is_psd(self.rho):
-                                raise ValueError("Density matrix must be positive semi-definite.")
-                        return
-                except TypeError:
-                    pass
+            if len(input_list) != 0 and isinstance(input_list[0], Iterable):
+
+                def get_row(
+                    item: Iterable[ExpressionOrSupportsComplex] | State | Expression | SupportsComplex,
+                ) -> list[ExpressionOrSupportsComplex]:
+                    if isinstance(item, Iterable):
+                        return list(item)
+                    raise TypeError("Every row of a matrix should be iterable.")
+
+                input_matrix: list[list[ExpressionOrSupportsComplex]] = [get_row(item) for item in input_list]
+                self.rho = np.array(input_matrix)
+                if not lv.is_qubitop(self.rho):
+                    raise ValueError("Cannot interpret the provided density matrix as a qubit operator.")
+                check_size_consistency(self.rho)
+                if self.rho.dtype != np.object_:
+                    if not lv.is_unit_trace(self.rho):
+                        raise ValueError("Density matrix must have unit trace.")
+                    if not lv.is_psd(self.rho):
+                        raise ValueError("Density matrix must be positive semi-definite.")
+                return
         statevec = Statevec(data, nqubit)
         # NOTE this works since np.outer flattens the inputs!
-        self.rho = np.outer(statevec.psi, statevec.psi.conj())
+        self.rho = outer(statevec.psi, statevec.psi.conj())
 
     @property
     def nqubit(self) -> int:
@@ -105,12 +111,36 @@ class DensityMatrix(State):
         """Return a string description."""
         return f"DensityMatrix object, with density matrix {self.rho} and shape {self.dims()}."
 
-    def add_nodes(self, nqubit, data) -> None:
-        """Add nodes to the density matrix."""
+    @override
+    def add_nodes(self, nqubit: int, data: Data) -> None:
+        r"""
+        Add nodes (qubits) to the density matrix and initialize them in a specified state.
+
+        Parameters
+        ----------
+        nqubit : int
+            The number of qubits to add to the density matrix.
+
+        data : Data, optional
+            The state in which to initialize the newly added nodes.
+
+            - If a single basic state is provided, all new nodes are initialized in that state.
+            - If a list of basic states is provided, it must match the length of ``nodes``, and
+              each node is initialized with its corresponding state.
+            - A single-qubit state vector will be broadcast to all nodes.
+            - A multi-qubit state vector of dimension :math:`2^n` initializes the new nodes jointly.
+            - A density matrix must have shape :math:`2^n \times 2^n`,
+              and is used to jointly initialize the new nodes.
+
+        Notes
+        -----
+        Previously existing nodes remain unchanged.
+        """
         dm_to_add = DensityMatrix(nqubit=nqubit, data=data)
         self.tensor(dm_to_add)
 
-    def evolve_single(self, op, i) -> None:
+    @override
+    def evolve_single(self, op: Matrix, i: int) -> None:
         """Single-qubit operation.
 
         Parameters
@@ -126,11 +156,12 @@ class DensityMatrix(State):
             raise ValueError("op must be 2*2 matrix.")
 
         rho_tensor = self.rho.reshape((2,) * self.nqubit * 2)
-        rho_tensor = np.tensordot(np.tensordot(op, rho_tensor, axes=(1, i)), op.conj().T, axes=(i + self.nqubit, 0))
+        rho_tensor = tensordot(tensordot(op, rho_tensor, axes=(1, i)), op.conj().T, axes=(i + self.nqubit, 0))
         rho_tensor = np.moveaxis(rho_tensor, (0, -1), (i, i + self.nqubit))
         self.rho = rho_tensor.reshape((2**self.nqubit, 2**self.nqubit))
 
-    def evolve(self, op: npt.NDArray, qargs: Collection[int]) -> None:
+    @override
+    def evolve(self, op: Matrix, qargs: Sequence[int]) -> None:
         """Multi-qubit operation.
 
         Args:
@@ -165,10 +196,10 @@ class DensityMatrix(State):
 
         rho_tensor = self.rho.reshape((2,) * self.nqubit * 2)
 
-        rho_tensor = np.tensordot(
-            np.tensordot(op_tensor, rho_tensor, axes=[tuple(nqb_op + i for i in range(len(qargs))), tuple(qargs)]),
+        rho_tensor = tensordot(
+            tensordot(op_tensor, rho_tensor, axes=(tuple(nqb_op + i for i in range(len(qargs))), tuple(qargs))),
             op.conj().T.reshape((2,) * 2 * nqb_op),
-            axes=[tuple(i + self.nqubit for i in qargs), tuple(i for i in range(len(qargs)))],
+            axes=(tuple(i + self.nqubit for i in qargs), tuple(i for i in range(len(qargs)))),
         )
         rho_tensor = np.moveaxis(
             rho_tensor,
@@ -177,7 +208,8 @@ class DensityMatrix(State):
         )
         self.rho = rho_tensor.reshape((2**self.nqubit, 2**self.nqubit))
 
-    def expectation_single(self, op: npt.NDArray, i: int) -> complex:
+    @override
+    def expectation_single(self, op: Matrix, loc: int) -> complex:
         """Return the expectation value of single-qubit operator.
 
         Args:
@@ -188,8 +220,8 @@ class DensityMatrix(State):
         -------
             complex: expectation value (real for hermitian ops!).
         """
-        if not (0 <= i < self.nqubit):
-            raise ValueError(f"Wrong target qubit {i}. Must between 0 and {self.nqubit - 1}.")
+        if not (0 <= loc < self.nqubit):
+            raise ValueError(f"Wrong target qubit {loc}. Must between 0 and {self.nqubit - 1}.")
 
         if op.shape != (2, 2):
             raise ValueError("op must be 2x2 matrix.")
@@ -197,11 +229,13 @@ class DensityMatrix(State):
         st1 = copy.copy(self)
         st1.normalize()
 
-        rho_tensor = st1.rho.reshape((2,) * st1.nqubit * 2)
-        rho_tensor = np.tensordot(op, rho_tensor, axes=[1, i])
-        rho_tensor = np.moveaxis(rho_tensor, 0, i)
+        nqubit = self.nqubit
+        rho_tensor: Matrix = st1.rho.reshape((2,) * nqubit * 2)
+        rho_tensor = tensordot(op, rho_tensor, axes=((1,), (loc,)))
+        rho_tensor = np.moveaxis(rho_tensor, 0, loc)
 
-        return np.trace(rho_tensor.reshape((2**self.nqubit, 2**self.nqubit)))
+        # complex() needed with mypy strict mode (no-any-return)
+        return complex(np.trace(rho_tensor.reshape((2**nqubit, 2**nqubit))))
 
     def dims(self) -> tuple[int, ...]:
         """Return the dimensions of the density matrix."""
@@ -219,7 +253,7 @@ class DensityMatrix(State):
         """
         if not isinstance(other, DensityMatrix):
             other = DensityMatrix(other)
-        self.rho = np.kron(self.rho, other.rho)
+        self.rho = kron(self.rho, other.rho)
 
     def cnot(self, edge: tuple[int, int]) -> None:
         """Apply CNOT gate to density matrix.
@@ -231,15 +265,16 @@ class DensityMatrix(State):
         """
         self.evolve(CNOT_TENSOR.reshape(4, 4), edge)
 
-    def swap(self, edge: tuple[int, int]) -> None:
+    @override
+    def swap(self, qubits: tuple[int, int]) -> None:
         """Swap qubits.
 
         Parameters
         ----------
-            edge : (int, int) or [int, int]
+            qubits : (int, int)
                 (control, target) qubits indices.
         """
-        self.evolve(SWAP_TENSOR.reshape(4, 4), edge)
+        self.evolve(SWAP_TENSOR.reshape(4, 4), qubits)
 
     def entangle(self, edge: tuple[int, int]) -> None:
         """Connect graph nodes.
@@ -253,11 +288,21 @@ class DensityMatrix(State):
 
     def normalize(self) -> None:
         """Normalize density matrix."""
-        self.rho /= np.trace(self.rho)
+        # Note that the following calls to `astype` are guaranteed to
+        # return the original NumPy array itself, since `copy=False` and
+        # the `dtype` matches. This is important because the array is
+        # then modified in place.
+        if self.rho.dtype == np.object_:
+            rho_o = self.rho.astype(np.object_, copy=False)
+            rho_o /= np.trace(rho_o)
+        else:
+            rho_c = self.rho.astype(np.complex128, copy=False)
+            rho_c /= np.trace(rho_c)
 
-    def remove_qubit(self, loc) -> None:
+    @override
+    def remove_qubit(self, qarg: int) -> None:
         """Remove a qubit."""
-        self.ptrace(loc)
+        self.ptrace(qarg)
         self.normalize()
 
     def ptrace(self, qargs: Collection[int] | int) -> None:
@@ -280,13 +325,12 @@ class DensityMatrix(State):
         rho_res = self.rho.reshape((2,) * n * 2)
         # ket, bra indices to trace out
         trace_axes = list(qargs) + [n + qarg for qarg in qargs]
-        rho_res = np.tensordot(
-            np.eye(2**qargs_num).reshape((2,) * qargs_num * 2), rho_res, axes=(list(range(2 * qargs_num)), trace_axes)
-        )
+        op: Matrix = np.eye(2**qargs_num).reshape((2,) * qargs_num * 2).astype(np.complex128)
+        rho_res = tensordot(op, rho_res, axes=(range(2 * qargs_num), trace_axes))
 
         self.rho = rho_res.reshape((2**nqubit_after, 2**nqubit_after))
 
-    def fidelity(self, statevec: Statevec) -> float:
+    def fidelity(self, statevec: Statevec) -> ExpressionOrFloat:
         """Calculate the fidelity against reference statevector.
 
         Parameters
@@ -294,13 +338,18 @@ class DensityMatrix(State):
             statevec : numpy array
                 statevector (flattened numpy array) to compare with
         """
-        return np.abs(statevec.transpose().conj() @ self.rho @ statevec)
+        result = vdot(statevec.psi, matmul(self.rho, statevec.psi))
+        if isinstance(result, Expression):
+            return result
+        assert math.isclose(result.imag, 0)
+        return result.real
 
-    def flatten(self) -> npt.NDArray:
+    def flatten(self) -> Matrix:
         """Return flattened density matrix."""
         return self.rho.flatten()
 
-    def apply_channel(self, channel: KrausChannel, qargs: Collection[int]) -> None:
+    @override
+    def apply_channel(self, channel: KrausChannel, qargs: Sequence[int]) -> None:
         """Apply a channel to a density matrix.
 
         Parameters
@@ -350,41 +399,8 @@ class DensityMatrix(State):
         return result
 
 
-class DensityMatrixBackend(Backend):
+@dataclass(frozen=True)
+class DensityMatrixBackend(DenseStateBackend[DensityMatrix]):
     """MBQC simulator with density matrix method."""
 
-    def __init__(self, **kwargs) -> None:
-        """Construct a density matrix backend."""
-        super().__init__(DensityMatrix(nqubit=0), **kwargs)
-
-    def apply_channel(self, channel: KrausChannel, qargs: Collection[int]) -> None:
-        """Apply channel to the state.
-
-        Parameters
-        ----------
-            qargs : list of ints. Target qubits
-        """
-        indices = [self.node_index.index(i) for i in qargs]
-        self.state.apply_channel(channel, indices)
-
-
-if sys.version_info >= (3, 10):
-    Data = (
-        states.State
-        | DensityMatrix
-        | Statevec
-        | Iterable[states.State]
-        | Iterable[ExpressionOrSupportsComplex]
-        | Iterable[Iterable[ExpressionOrSupportsComplex]]
-    )
-else:
-    from typing import Union
-
-    Data = Union[
-        states.State,
-        DensityMatrix,
-        Statevec,
-        Iterable[states.State],
-        Iterable[ExpressionOrSupportsComplex],
-        Iterable[Iterable[ExpressionOrSupportsComplex]],
-    ]
+    state: DensityMatrix = dataclasses.field(init=False, default_factory=lambda: DensityMatrix(nqubit=0))